python-service/requirements.txt
fastapi
uvicorn
opencv-python
pillow
numpy
python-multipart
python-dotenv
APScheduler
pymongo
requests
groq
